fix(agent): close last two Opus follow-ups on #284 T1 egress guard
Opus approved PR #458 with two small cleanup items:

1. egress_guard.py's RESOLUTION_FAILED branch (OSError/socket.gaierror)
   was the module's only uncovered line. Added a test that monkeypatches
   socket.getaddrinfo to raise gaierror and asserts the EgressBlockReason.
   New DNS-resolution tests split into test_egress_dns_resolution.py to
   keep test_egress_guard_classification.py under the 200-line cap.

2. egress_transport.py imported egress_guard's private _default_resolve
   across a module boundary. Renamed to the public default_resolve in
   egress_guard.py; egress_transport.py and the DNS test now import/refer
   to the public name.

Mutation gates re-verified after the rename (TOCTOU + both halves of the
dual is_global/deny-flag condition all still fail correctly when disabled).
make lint / make typecheck / make test green: 1434 passed, 88.67% coverage
(floor 87%); egress_guard.py, egress_transport.py, egress_errors.py all at
100% line coverage.

Refs #284
